fix: Selecting permissions for a location
Commit content:
Change the permission request to click the button (initially open the app, the system will ask for permission).
When the user clicks the button, if he does not have permission, he will ask the user to grant permission.
If previously selecting only this time, the user is forced to grant permission when clicking the button when re-entering the app.
If you are denied and continue to click the button, the system will continue to ask for permission. If denied, the system displays a dialog for the user to grant permission manually.If the user denies the permission permanently (deny 2 times with Android 11+ and select Don't show again with Android 9 -), when re-entering the app, the system will display a dialog for the user to grant the permission manually.
In the Android 15 version, there will be an exact location and relative location option when requesting location permissions.